+++ askbot/views/writers.py
@@ -702,6 +702,7 @@ def post_comments(request):#generic ajax handler to load comments to an object
 @decorators.ajax_only
 #@decorators.check_spam('comment')
 def edit_comment(request):
+
     if request.user.is_anonymous():
         raise exceptions.PermissionDenied(_('Sorry, anonymous users cannot edit comments'))